Rename transaction-utils file
